# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Tiles Stock Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Depot {
  id: number;
  name: string;
  /** @nullable */
  location?: string | null;
  createdAt: string;
  /** @nullable */
  lastUploadAt?: string | null;
  itemCount: number;
}

export interface DepotInput {
  /** @minLength 1 */
  name: string;
  location?: string;
}

export interface StockItem {
  id: number;
  depotId: number;
  depotName: string;
  tileName: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  design?: string | null;
  /** @nullable */
  finish?: string | null;
  /** @nullable */
  boxCount?: number | null;
  /** @nullable */
  pcsCount?: number | null;
  /** @nullable */
  stockDate?: string | null;
  uploadId: number;
  updatedAt: string;
  hasImage: boolean;
}

export interface StockPage {
  items: StockItem[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface DepotSummary {
  depotId: number;
  depotName: string;
  totalItems: number;
  totalBoxes: number;
  /** @nullable */
  lastUploadAt?: string | null;
  /** @nullable */
  stockDate?: string | null;
}

export interface StockFilters {
  brands: string[];
  sizes: string[];
  finishes: string[];
}

export type UploadStatus = typeof UploadStatus[keyof typeof UploadStatus];


export const UploadStatus = {
  pending: 'pending',
  processing: 'processing',
  done: 'done',
  failed: 'failed',
} as const;

export interface Upload {
  id: number;
  depotId: number;
  depotName: string;
  filename: string;
  status: UploadStatus;
  /** @nullable */
  itemsExtracted?: number | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  stockDate?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface UploadInput {
  depotId: number;
  stockDate?: string;
  file: Blob;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  role: string;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};

export type UpdateDepotBody = {
  name?: string;
  /** @nullable */
  location?: string | null;
};

export type ListStockParams = {
/**
 * Search by tile name, brand, or design
 */
search?: string;
/**
 * Filter by depot ID
 * @nullable
 */
depotId?: number | null;
/**
 * Filter by brand name
 * @nullable
 */
brand?: string | null;
/**
 * Filter by tile size (e.g. 1200X600)
 * @nullable
 */
size?: string | null;
/**
 * Filter by finish (GLOSSY, MATT, etc.)
 * @nullable
 */
finish?: string | null;
page?: number;
limit?: number;
};

export type GetStockFiltersParams = {
/**
 * @nullable
 */
depotId?: number | null;
};

export type ListUploadsParams = {
/**
 * @nullable
 */
depotId?: number | null;
limit?: number;
};