# lib/api-zod/src/generated/types/mobileTokenExchangeSuccess.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Tiles Stock Management API
 * OpenAPI spec version: 0.1.0
 */

export interface MobileTokenExchangeSuccess {
  token: string;
}